pyproject.toml
[tool.poetry]
name = "notifier"
version = "0.1.0"
description = ""
authors = ["Magnus <magnus@wemade.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
pika = "^1.3.2"
python-dotenv = "^1.0.1"
sendgrid = "^6.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
